Rt to float, ADAPChromatogram not implementing Feature, modular 3D visualization

+++ src/main/java/io/github/mzmine/modules/io/projectload/version_2_5/RawDataFileOpenHandler_2_5.java
@@ -58,7 +58,7 @@ public class RawDataFileOpenHandler_2_5 extends DefaultHandler implements RawDat
   private int numberOfFragments;
   private double precursorMZ;
   private int precursorCharge;
-  private double retentionTime;
+  private float retentionTime;
   private double mobility;
   private int dataPointsNumber;
   private int fragmentCount;
@@ -222,7 +222,7 @@ public class RawDataFileOpenHandler_2_5 extends DefaultHandler implements RawDat
     if (qName.equals(RawDataElementName_2_5.RETENTION_TIME.getElementName())) {
       // Before MZmine.6 retention time was saved in seconds, but now we
       // use minutes, so we need to divide by 60
-      retentionTime = Double.parseDouble(getTextOfElement()) / 60d;
+      retentionTime = (float) (Double.parseDouble(getTextOfElement()) / 60d);
     }
 
     /* TODO:
