+++ src/main/java/io/github/mzmine/parameters/parametertypes/selectors/ScanSelection.java
@@ -36,7 +36,7 @@ public class ScanSelection {
 
   private final Range<Integer> scanNumberRange;
   private Integer baseFilteringInteger;
-  private final Range<Double> scanRTRange;
+  private final Range<Float> scanRTRange;
   private final PolarityType polarity;
   private final MassSpectrumType spectrumType;
   private final Integer msLevel;
@@ -50,12 +50,12 @@ public class ScanSelection {
     this(null, null, null, null, null, msLevel, null);
   }
 
-  public ScanSelection(Range<Double> scanRTRange, int msLevel) {
+  public ScanSelection(Range<Float> scanRTRange, int msLevel) {
     this(null, null, scanRTRange, null, null, msLevel, null);
   }
 
   public ScanSelection(Range<Integer> scanNumberRange, Integer baseFilteringInteger,
-      Range<Double> scanRTRange, PolarityType polarity, MassSpectrumType spectrumType,
+      Range<Float> scanRTRange, PolarityType polarity, MassSpectrumType spectrumType,
       Integer msLevel, String scanDefinition) {
     this.scanNumberRange = scanNumberRange;
     this.baseFilteringInteger = baseFilteringInteger;
@@ -74,7 +74,7 @@ public class ScanSelection {
     return baseFilteringInteger;
   }
 
-  public Range<Double> getScanRTRange() {
+  public Range<Float> getScanRTRange() {
     return scanRTRange;
   }
